camerawrapper: Fixup camera-mode logic
Previously, this set camera-mode to 1 (enableZSL in camera app Java)
when recording-hint is not present at all.

It also set it to 0 when recording.

However, in the camera app, if enableZSL is set in the camera overlay,
camera-mode is always 1.

So fixup setparams to always set camera-mode to 1.

This should fix the issue where preview and pictures are initially dark
until the user switches to video and back.

Change-Id: Ie9c2e0328b8a5ab2943b87689b0cdfd1d306e2da