Update patch set 4
Patch Set 4:

Not sure if there is something like a "default" layout for a device tree. Think we have never thought about this and the layout has been copy/pasted from device to device :)

This layout has been somehow settled which doesnt mean that it is the best. Some comments

The keylayout stuff should be under configs/keylayout - thats the way it is on most other devices now

I like the idea of putting the qcom stuff in a separate folder even if configs also contains scripts. Also having a folder for all ramdisk files would make sense - naming it root like in the build is ok for me

Patch-set: 4